requirements.txt
requests
cohere
python-dotenv
google-generativeai